requirements.txt
ops~=1.1.0
redis~=3.5.3

PyYAML~=5.3.1
charmhelpers~=0.20.20